✨ handle `label` & `-label` filters
